refactor: extract backend specific logic to backend folders
